+++ EXE201.Server/Program.cs
@@ -62,6 +62,8 @@ namespace EXE201.Server
             builder.Services.AddHostedService<EXE201.Server.Services.BookingExpiryWorker>();
             builder.Services.AddHttpClient<EXE201.Server.Services.IGeminiModerationService, EXE201.Server.Services.GeminiModerationService>();
             builder.Services.AddHttpClient<EXE201.Server.Services.IGeminiChatbotService, EXE201.Server.Services.GeminiChatbotService>();
+            builder.Services.AddScoped<EXE201.Server.Repositories.IWalletRepository, EXE201.Server.Repositories.WalletRepository>();
+            builder.Services.AddScoped<EXE201.Server.Services.IWalletService, EXE201.Server.Services.WalletService>();
 
             // Register payOS
             builder.Services.AddSingleton(new PayOS.PayOSClient(